Move the PDF template classes into their own package
Fixes: SE-8489
